+++ src/Ampersand/Commands/Devoutput.hs
@@ -13,6 +13,7 @@ import Ampersand.Basics
 import Ampersand.FSpec
 import Ampersand.Misc.HasClasses
 import Ampersand.Output.FSpec2SQL
+import Ampersand.Output.FSpec2Turtle
 import qualified RIO.Text as T
 import RIO.Time
 import System.Directory
@@ -24,8 +25,15 @@ devoutput ::
   FSpec ->
   RIO env ()
 devoutput fSpec = do
-  doGenHaskell fSpec
-  doGenSQLdump fSpec
+  -- For the time being, do not output Haskell and SQL dump files.
+  -- This is done in a way that there is no dead code. Feel free to
+  -- move the doGen... functions down if you want to generate these files.
+  case originalContext fSpec of
+    Nothing -> do
+      doGenHaskell fSpec
+      doGenSQLdump fSpec
+    Just _ -> mempty
+  writeTurtle fSpec
 
 doGenHaskell ::
   (HasDirOutput env, HasFSpecGenOpts env, HasLogFunc env) =>